Update: - README updates - cleanup
