+++ app/src/main/java/com/neudesic/myapplication/di/NetworkingModule.kt
@@ -42,9 +42,6 @@ object NetworkingModule {
         return "https://someotherurl.com"
     }
 
-    // NOTES: BaseUrl > provideRetrofitClient
-    //        provideOkHttpClient > provideRetrofitClient
-    //        provideRetrofitClient > provideRestApiService
     @Provides
     fun provideOkHttpClient(): OkHttpClient {
         return OkHttpClient()
